requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# File handling
python-multipart==0.0.6
openpyxl==3.1.2

# Templates
jinja2==3.1.3

# Utils
python-dotenv==1.0.0
apscheduler==3.10.4

# Production server
gunicorn==21.2.0
